Fixed FileError handling. Deprecated to use DomError and a name attribute.
Notes:
Basically file ops can throw FileError (currently), which is a subclass of
DomError. FileError is considered deprecated and .code is no longer used, rather
.name in DomError.

Wrapping this, due to the implementation of the file ops, the DomError actually
thrown is currently in the form of a JsObject, hence the uncertainty as to
whether it will be JsObject or DomError (i.e. whether it is properly proxied to
a Dart object).

Minor filesystem fix.

BUG=
R=koz@chromium.org

Review URL: https://codereview.appspot.com/51830044